Feature: adicionando testes nas services

+++ Sprint 2/JavaSpringBoot/projetoMima/src/main/java/school/sptech/projetoMima/service/auxiliares/TamanhoService.java
@@ -4,6 +4,8 @@ import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import school.sptech.projetoMima.entity.item.Tamanho;
 import school.sptech.projetoMima.exception.Item.Auxiliares.TamanhoDuplicadoException;
+import school.sptech.projetoMima.exception.Item.Auxiliares.TamanhoListaVaziaException;
+import school.sptech.projetoMima.exception.Item.Auxiliares.TamanhoNaoEncontradoException;
 import school.sptech.projetoMima.repository.auxiliares.TamanhoRepository;
 
 import java.util.List;
@@ -25,14 +27,22 @@ public class TamanhoService {
 
     public Tamanho buscarPorId(Integer id) {
         return repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Tamanho não encontrado"));
+                .orElseThrow(() -> new TamanhoNaoEncontradoException("Tamanho não encontrado"));
     }
 
     public void deletar(Integer id) {
+        if(!repository.existsById(id)) {
+            throw new TamanhoNaoEncontradoException("esse tamanho não existe na lista!");
+        }
+
         repository.deleteById(id);
     }
 
     public List<Tamanho> listar() {
+        if (repository.findAll().isEmpty()) {
+            throw new TamanhoListaVaziaException("não há tamanhos cadastrados");
+        }
+
         return repository.findAll();
     }
 }
